# specs/openapi.yaml
openapi: 3.1.0
info:
  title: bracket-creator API
  description: |
    Web API for generating Kendo tournament brackets.

    This spec covers two servers:
    - **Bracket Generator** (`serve` command): Stateless, form-based bracket generation.
    - **Mobile/Admin API** (`mobile-app` command): Stateful REST API for live tournament management.

    The Mobile/Admin API routes under `/api/` (excluding `/api/viewer/`) require the
    `X-Tournament-Password` header for authentication.
  version: 2.1.0
servers:
  - url: http://localhost:8080
    description: Bracket Generator (serve command)
  - url: http://localhost:8081
    description: Mobile/Admin API (mobile-app command)
...
paths:
...
  /api/competitions/{id}/participants:
    get:
      tags: [participants]
      summary: Get participants
      security:
        - TournamentPassword: []
      parameters:
        - $ref: '#/components/parameters/CompetitionId'
      responses:
        '200':
          description: List of players.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Player'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [participants]
      summary: Set participants (JSON)
      description: Replaces the current participants list for the competition with the provided JSON.
      security:
        - TournamentPassword: []
      parameters:
        - $ref: '#/components/parameters/CompetitionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                players:
                  type: array
                  items:
                    $ref: '#/components/schemas/Participant'
      responses:
        '200':
          description: Participants successfully updated.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Player'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/competitions/{id}/seeds:
    get:
      tags: [participants]
      summary: Get seed assignments
      security:
        - TournamentPassword: []
      parameters:
        - $ref: '#/components/parameters/CompetitionId'
      responses:
        '200':
          description: Current seed assignments.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SeedAssignment'
    put:
      tags: [participants]
      summary: Save seed assignments
      security:
        - TournamentPassword: []
      parameters:
        - $ref: '#/components/parameters/CompetitionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/SeedAssignment'
      responses:
        '200':
          description: Seeds saved.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SeedAssignment'

  # ── Matches (Admin, requires auth) ─────────────────────────────────────────
...
components:
  securitySchemes:
    TournamentPassword:
      type: apiKey
      in: header
      name: X-Tournament-Password
      description: |
        Tournament admin password. Required for all `/api/` routes except
        `/api/viewer/`, `/api/events`, and the bootstrap PUT/POST
        `/api/tournament` while no tournament has been configured yet.

        Header value must exactly equal the stored tournament password
        (exact-string match; no trimming applied to either side).

        Defense-in-depth: if the stored tournament password is somehow
        empty (e.g. operator manually edited `tournament.md` to land
        an empty `password:` field), AuthMiddleware returns 403
        (`tournament misconfigured: password is not set`) for ALL
        requests rather than letting the `header != stored` check
        vacuously pass for empty header.

        Recovery from this state requires **out-of-band repair**: the
        bootstrap PUT/POST `/api/tournament` exception only applies
        when the loaded record is `nil` or matches the literal
        `Name == "New Tournament" && Password == ""` default — a
        real-named record with empty Password is locked out by the 403
        fail-closed branch. Operator must stop the server, edit
        `tournament.md` to either delete it (forces fresh bootstrap)
        or set a non-empty `password:` field in the YAML front-matter,
        then restart.
  parameters:
    CompetitionId:
      in: path
      name: id
      required: true
      schema:
        type: string
        pattern: '^[a-zA-Z0-9][a-zA-Z0-9_-]*$'
        maxLength: 64
      description: |
        Competition slug/ID (e.g. "mens-individual"). Must start with an
        alphanumeric character, contain only alphanumerics / hyphens /
        underscores, and be at most 64 chars. Validated server-side via
        `requireValidCompID`; invalid IDs return 400
        (`competition ID contains invalid characters` / `competition ID too long` /
        `competition ID cannot be empty`). The validation is a path-traversal
        defense: without it, an `id` like `../../../etc/passwd` would resolve
        through `filepath.Clean(filepath.Join(...))` and escape the data
        directory.
...
  responses:
    BadRequest:
      description: Invalid input.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    Player:
      type: object
      properties:
        id:
          type: string
          maxLength: 64
        name:
          type: string
          maxLength: 100
        displayName:
          type: string
          maxLength: 50
          description: Display name shown on the bracket and Zekken. Defaults to a sanitized form of `name` when omitted.
        dojo:
          type: string
          maxLength: 100
        metadata:
          type: array
          maxItems: 16
          items:
            type: string
            maxLength: 200
        tag:
          type: string
          maxLength: 200
        seed:
          type: integer
          description: Seed rank (0 = unseeded).
        number:
          type: string
          description: Match number prefix + sequence (e.g. "K1"). Only present when number-prefix is set.
    Participant:
      type: object
      properties:
        id:
          type: string
          maxLength: 64
        name:
          type: string
          maxLength: 100
        displayName:
          type: string
          maxLength: 50
        dojo:
          type: string
          maxLength: 100
        metadata:
          type: array
          maxItems: 16
          items:
            type: string
            maxLength: 200
        tag:
          type: string
          maxLength: 200
    SeedAssignment:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        seedRank:
          type: integer
          description: 1 = top seed.
...